pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdb-spacecraft-simulator"
version = "0.1.0"
description = "Schedule-driven spacecraft mission simulation stack for LEO/VLEO constellations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PDB Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "sgp4>=2.22",
    "pyproj>=3.5.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
]
validation = [
    "jinja2>=3.0.0",
]
aerie = [
    "requests>=2.28.0",
]
basilisk = [
    "Basilisk>=2.0.0",
]
medium = [
    "Basilisk>=2.0.0",
    "scipy>=1.10.0",
]
mcp = [
    "mcp>=0.9.0",
]
e2e = [
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "requests>=2.28.0",
]
modelgen = [
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
simrun = "cli.simrun:main"
mcp-server = "sim_mcp.server:main"
modelgen = "tools.modelgen.cli:modelgen"

[tool.setuptools.packages.find]
include = ["sim*", "cli*", "sim_mcp*", "tools*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v"
asyncio_mode = "auto"
markers = [
    "e2e: mark test as end-to-end test requiring Aerie",
    "ete: mark test as end-to-end validation test",
    "ete_smoke: ETE smoke tests (<60s)",
    "ete_tier_a: ETE Tier A tests (<300s)",
    "ete_tier_b: ETE Tier B tests (<1800s, nightly)",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
